socklen_t is "int" in the Winsocks API
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4496 d7cf8633-e32d-0410-b094-e92efae38249

+++ lhnet.c
@@ -460,7 +460,7 @@ lhnetsocket_t *LHNET_OpenSocket_Connectionless(lhnetaddress_t *address)
 #endif
 					{
 #ifdef WIN32
-						size_t namelen;
+						int namelen;
 #else
 						socklen_t namelen;
 #endif
